Split inductor lowerings by category
