+++ torch/_inductor/lowering.py
@@ -5072,2210 +5072,15 @@ def constant_boundary_condition(
     return load
 
 
-def pooling_size(x, i, kernel_size, stride, padding, ceil_mode, *, dilation=None):
-    if dilation is None:
-        dilation = [1] * len(padding)
 
-    x_out = FloorDiv(
-        x + 2 * padding[i] - dilation[i] * (kernel_size[i] - 1) + (stride[i] - 1),
-        stride[i],
-    )
 
-    if ceil_mode:
-        x_alt = FloorDiv(
-            x
-            + 2 * padding[i]
-            - dilation[i] * (kernel_size[i] - 1)
-            + 2 * (stride[i] - 1),
-            stride[i],
-        )
-        if V.graph.sizevars.guard_or_false(
-            sympy.Ge((x_alt - 1) * stride[i] - x - padding[i], 0)
-        ):
-            # Sliding windows must start within the input or left padding
-            x_alt -= 1  # type: ignore[assignment]
-        if V.graph.sizevars.guard_or_false(sympy.Eq(x_out, x_alt)):
-            # ceil mode is actually a no-op, lets guard on that
-            ceil_mode = False
-        else:
-            x_out = x_alt
-    return x_out, ceil_mode
 
 
-def should_fallback_max_pool_with_indices(kernel_size, *, n_dim):
-    kernel_size = pad_listlike(kernel_size, n_dim)
-    window_size = functools.reduce(operator.mul, kernel_size)
-    return window_size > 25
 
 
-def max_pool_checks(
-    x, kernel_size, stride, padding, dilation, n_dim, *, assert_fallback=None
-):
-    if padding == 0:
-        padding = [0] * n_dim
-    if dilation == 1:
-        dilation = [1] * n_dim
-    if not stride:
-        stride = kernel_size
 
-    kernel_size = pad_listlike(kernel_size, n_dim)
-    stride = pad_listlike(stride, n_dim)
-    padding = pad_listlike(padding, n_dim)
-    dilation = pad_listlike(dilation, n_dim)
 
-    assert isinstance(x, TensorBox)
-    assert len(kernel_size) == n_dim
-    assert len(stride) == n_dim
-    assert len(padding) == n_dim
-    assert len(dilation) == n_dim
-    assert len(x.get_size()) in (n_dim + 1, n_dim + 2)
 
-    use_fallback = should_fallback_max_pool_with_indices(kernel_size, n_dim=n_dim)
-    if assert_fallback is not None:
-        assert use_fallback == assert_fallback
-
-    return kernel_size, stride, padding, dilation, use_fallback
-
-
-def _max_pool_with_offsets(
-    x,
-    kernel_size,
-    stride,
-    padding,
-    dilation,
-    ceil_mode,
-    *,
-    n_dim,
-):
-    x.realize_hint()
-    batch = x.shape[:-n_dim]
-    dhw = x.shape[-n_dim:]
-
-    dhw_out, ceil_mode = zip(
-        *[
-            pooling_size(
-                dhw[d], d, kernel_size, stride, padding, ceil_mode, dilation=dilation
-            )
-            for d in range(n_dim)
-        ]
-    )
-
-    dtype = x.dtype
-    min_value = (
-        False
-        if dtype is torch.bool
-        else (float("-inf") if dtype.is_floating_point else torch.iinfo(dtype).min)
-    )
-
-    new_size = list(batch) + list(dhw_out)
-    if any(padding) or any(ceil_mode) or any(d > 1 for d in dilation):
-        x_loader = constant_boundary_condition(x, min_value, dim=n_dim)
-    else:
-        x_loader = x.make_loader()
-
-    def fn_inner(idx, reduction_idx):
-        prefix = idx[:-n_dim]
-        bh = idx[-n_dim:]
-        ih = [
-            (bh[i] * stride[i]) + (reduction_idx[i] * dilation[i]) - padding[i]
-            for i in range(n_dim)
-        ]
-        return x_loader([*prefix, *ih])
-
-    result = Reduction.create(
-        reduction_type="max",
-        input_node=x,
-        device=x.get_device(),
-        dst_dtype=dtype,
-        src_dtype=dtype,
-        inner_fn=fn_inner,
-        ranges=new_size,
-        reduction_ranges=kernel_size,
-    )
-    offsets = Reduction.create(
-        reduction_type="argmax",
-        input_node=x,
-        device=x.get_device(),
-        dst_dtype=torch.int64,
-        src_dtype=dtype,
-        inner_fn=fn_inner,
-        ranges=new_size,
-        reduction_ranges=kernel_size,
-    )
-    if isinstance(result.data.data, Reduction):  # type: ignore[attr-defined, union-attr]
-        # Only realize if reduction isn't unrolled
-        result.realize()
-    if isinstance(offsets.data.data, Reduction):  # type: ignore[attr-defined, union-attr]
-        # Only realize if reduction isn't unrolled
-        offsets.realize()
-
-    return result, offsets
-
-
-@register_lowering(prims._low_memory_max_pool_with_offsets, type_promotion_kind=None)
-def _low_memory_max_pool_with_offsets(
-    x,
-    kernel_size,
-    stride,
-    padding,
-    dilation,
-    ceil_mode=False,
-):
-    n_dim = len(kernel_size)
-
-    # assert we are not on a fallback path, the inductor decomp should have guaranteed this
-    kernel_size, stride, padding, dilation, _ = max_pool_checks(
-        x,
-        kernel_size,
-        stride,
-        padding,
-        dilation,
-        n_dim,
-        assert_fallback=False,
-    )
-
-    with config.patch(unroll_reductions_threshold=25):
-        result, offsets = _max_pool_with_offsets(
-            x,
-            kernel_size,
-            stride,
-            padding,
-            dilation,
-            ceil_mode,
-            n_dim=n_dim,
-        )
-        return result, to_dtype(offsets, torch.int8)
-
-
-def _pool_offsets_to_indices(
-    offsets: TensorBox,
-    kernel_size: Sequence[int | torch.SymInt],
-    input_size: Sequence[int | torch.SymInt],
-    increments_to_index: Callable[
-        [Sequence[int | torch.SymInt], Sequence[int | torch.SymInt]],
-        torch._inductor.virtualized.OpsValue,
-    ],
-) -> TensorBox:
-    n_dim = len(kernel_size)
-    offsets_loader = offsets.make_loader()
-    window_size = sympy.sympify(functools.reduce(operator.mul, kernel_size))
-
-    def offsets_to_indices(idx):
-        offset = offsets_loader(idx)
-        offset_sympy = ops.indirect_indexing(offset, window_size)
-        reduction_idx = inductor_prims._flattened_index_to_nd(offset_sympy, kernel_size)
-        idhw = increments_to_index(idx, reduction_idx)
-        return ops.index_expr(
-            inductor_prims._flatten_index(idhw, input_size[-n_dim:]), torch.int64
-        )
-
-    indices = Pointwise.create(
-        device=offsets.get_device(),
-        dtype=torch.int64,
-        inner_fn=offsets_to_indices,
-        ranges=offsets.get_size(),
-    )
-    return indices
-
-
-@register_lowering(
-    prims._low_memory_max_pool_offsets_to_indices, type_promotion_kind=None
-)
-def _low_memory_max_pool_offsets_to_indices(
-    offsets, kernel_size, input_size, stride, padding, dilation
-):
-    # TODO: Generalize to other max pooling flavors
-    n_dim = len(kernel_size)
-
-    def increments_to_index(idx, reduction_idx):
-        bh = idx[-n_dim:]
-        return [
-            (bh[i] * stride[i]) + (reduction_idx[i] * dilation[i]) - padding[i]
-            for i in range(n_dim)
-        ]
-
-    return _pool_offsets_to_indices(
-        offsets, kernel_size, input_size, increments_to_index
-    )
-
-
-def _max_pool_with_indices(
-    x,
-    kernel_size,
-    stride,
-    padding,
-    dilation,
-    ceil_mode,
-    n_dim,
-):
-    kernel_size, stride, padding, dilation, _ = max_pool_checks(
-        x, kernel_size, stride, padding, dilation, n_dim=n_dim
-    )
-
-    out, offsets = _max_pool_with_offsets(
-        x, kernel_size, stride, padding, dilation, ceil_mode, n_dim=n_dim
-    )
-
-    indices = _low_memory_max_pool_offsets_to_indices(
-        offsets,
-        kernel_size,
-        x.shape[-n_dim:],
-        stride,
-        padding,
-        dilation,
-    )
-
-    return out, indices
-
-
-# Fallback when we do not decompose to the low-memory path.
-@register_lowering(aten.max_pool2d_with_indices, type_promotion_kind=None)
-def max_pool2d_with_indices(
-    x,
-    kernel_size,
-    stride=None,
-    padding=0,
-    dilation=1,
-    ceil_mode=False,
-):
-    return _max_pool_with_indices(
-        x, kernel_size, stride, padding, dilation, ceil_mode, n_dim=2
-    )
-
-
-# Fallback when we do not decompose to the low-memory path.
-@register_lowering(aten.max_pool3d_with_indices, type_promotion_kind=None)
-def max_pool3d_with_indices(
-    x,
-    kernel_size,
-    stride=None,
-    padding=0,
-    dilation=1,
-    ceil_mode=False,
-):
-    return _max_pool_with_indices(
-        x, kernel_size, stride, padding, dilation, ceil_mode, n_dim=3
-    )
-
-
-fallback_max_pool2d_with_indices_backward = fallback_handler(
-    aten.max_pool2d_with_indices_backward.default,
-    add_to_fallback_set=False,
-)
-
-
-@register_lowering(aten.max_pool2d_with_indices_backward, type_promotion_kind=None)
-def max_pool2d_with_indices_backward(
-    grad_output, x, kernel_size, stride, padding, dilation, ceil_mode, indices
-):
-    if padding == 0:
-        padding = [0, 0]
-    if dilation == 1:
-        dilation = [1, 1]
-    if not stride:
-        stride = kernel_size
-
-    assert isinstance(x, TensorBox)
-    assert len(kernel_size) == 2
-    assert len(stride) == 2
-    assert len(padding) == 2
-    assert len(dilation) == 2
-    assert len(x.get_size()) in (3, 4)
-
-    # we will read this many times, so make sure it is computed
-    grad_output.realize_hint()
-    gO_stride = grad_output.maybe_get_stride()
-    x_stride: Sequence[Any] | None
-    if isinstance(x, TensorBox) and isinstance(x.data.data, Pointwise):  # type: ignore[attr-defined]
-        data = x.data.data  # type: ignore[attr-defined]
-        device = data.get_device()
-        assert device is not None
-        x_buffer = ir.ComputedBuffer(
-            name=None,
-            layout=ir.FlexibleLayout(
-                device=device,
-                dtype=data.get_dtype(),
-                size=data.get_size(),
-            ),
-            data=data,
-        )
-        x_buffer.decide_layout()
-        x_stride = x_buffer.get_stride()
-    else:
-        x_stride = x.maybe_get_stride()
-
-    is_channels_last = (x_stride is not None and x_stride[1] == 1) or (
-        gO_stride is not None and gO_stride[1] == 1
-    )
-    if any(d != 1 for d in dilation):
-        # dilation NYI
-        return fallback_max_pool2d_with_indices_backward(
-            grad_output, x, kernel_size, stride, padding, dilation, ceil_mode, indices
-        )
-
-    *_batch, _height, width = x.get_size()
-    *_, pooled_height, pooled_width = grad_output.get_size()
-
-    indices_loader = indices.make_loader()
-    grad_loader = grad_output.make_loader()
-    new_size = list(x.get_size())
-
-    h_window_size = max(
-        max(FloorDiv(h, stride[0]) - max(0, FloorDiv(h - kernel_size[0], stride[0])), 1)
-        for h in range(kernel_size[0] * 2)
-    )
-    w_window_size = max(
-        max(FloorDiv(w, stride[1]) - max(0, FloorDiv(w - kernel_size[1], stride[1])), 1)
-        for w in range(kernel_size[1] * 2)
-    )
-
-    window_size = h_window_size * w_window_size
-
-    if window_size > 25:
-        # Kernel size too big. Results in hard-to-optimize Triton code. Use fallback.
-        return fallback_max_pool2d_with_indices_backward(
-            grad_output, x, kernel_size, stride, padding, dilation, ceil_mode, indices
-        )
-
-    indices_size = indices.get_size()
-
-    def fn(idx):
-        *prefix, h, w = idx
-        index_test = ops.index_expr(h * width + w, torch.int32)
-        h = h + padding[0]
-        w = w + padding[1]
-        phstart = ops.index_expr(
-            FloorDiv(h - kernel_size[0] + stride[0], stride[0]), torch.int32
-        )
-        pwstart = ops.index_expr(
-            FloorDiv(w - kernel_size[1] + stride[1], stride[1]), torch.int32
-        )
-        phend = ops.index_expr(FloorDiv(h, stride[0]) + 1, torch.int32)
-        pwend = ops.index_expr(FloorDiv(w, stride[1]) + 1, torch.int32)
-
-        phstart = ops.maximum(phstart, ops.constant(0, torch.int32))
-        pwstart = ops.maximum(pwstart, ops.constant(0, torch.int32))
-        phend = ops.minimum(phend, ops.index_expr(pooled_height, torch.int32))
-        pwend = ops.minimum(pwend, ops.index_expr(pooled_width, torch.int32))
-
-        gradient = None
-        for ph_ in range(h_window_size):
-            for pw_ in range(w_window_size):
-                ph = ops.add(phstart, ops.constant(ph_, torch.int32))
-                pw = ops.add(pwstart, ops.constant(pw_, torch.int32))
-                grad_index = [
-                    *prefix,
-                    ops.indirect_indexing(
-                        ops.minimum(ph, ops.sub(phend, ops.constant(1, torch.int32))),
-                        indices_size[-2],
-                        check=False,
-                    ),
-                    ops.indirect_indexing(
-                        ops.minimum(pw, ops.sub(pwend, ops.constant(1, torch.int32))),
-                        indices_size[-1],
-                        check=False,
-                    ),
-                ]
-
-                index_actual = indices_loader(grad_index)
-                grad_part = grad_loader(grad_index)
-                check = ops.eq(index_actual, index_test)
-
-                if gradient is None:
-                    # don't need mask for 0, 0
-                    gradient = ops.where(
-                        check, grad_part, ops.constant(0.0, torch.float32)
-                    )
-                else:
-                    mask = ops.and_(
-                        ops.and_(
-                            ops.lt(ph, phend),
-                            ops.lt(pw, pwend),
-                        ),
-                        check,
-                    )
-                    gradient = ops.where(mask, ops.add(gradient, grad_part), gradient)
-        assert gradient is not None
-        return gradient
-
-    out = Pointwise.create(
-        device=grad_output.get_device(),
-        dtype=grad_output.get_dtype(),
-        inner_fn=fn,
-        ranges=new_size,
-    )
-    if is_channels_last:
-        return ir.ExternKernel.require_channels_last(out)
-    else:
-        return out
-
-
-def pad_adaptive_loader(x, pad_val=0.0):
-    x_loader = x.make_loader()
-
-    def load(prefix, increments, start_indices, end_indices):
-        ih, iw = increments
-        h_start_index, w_start_index = start_indices
-        h_end_index, w_end_index = end_indices
-
-        mask = ops.and_(
-            ops.lt(
-                ops.index_expr(h_start_index + ih, torch.int64),
-                ops.index_expr(h_end_index, torch.int64),
-            ),
-            ops.lt(
-                ops.index_expr(w_start_index + iw, torch.int64),
-                ops.index_expr(w_end_index, torch.int64),
-            ),
-        )
-
-        return ops.masked(
-            mask,
-            lambda: x_loader([*prefix, h_start_index + ih, w_start_index + iw]),
-            pad_val,
-        )
-
-    return load
-
-
-def compute_indices_adaptive_pooling(start_index, end_index, h_in, w_in, h_out, w_out):
-    h_start_index = functools.partial(start_index, out_dim=h_out, inp_dim=h_in)
-    h_end_index = functools.partial(end_index, out_dim=h_out, inp_dim=h_in)
-
-    w_start_index = functools.partial(start_index, out_dim=w_out, inp_dim=w_in)
-    w_end_index = functools.partial(end_index, out_dim=w_out, inp_dim=w_in)
-
-    return h_start_index, h_end_index, w_start_index, w_end_index
-
-
-def _adaptive_pooling_fn(
-    start_index, end_index, kernel_maxes, in_sizes, out_sizes, pooling_fn
-):
-    h_in, w_in = in_sizes
-    h_out, w_out = out_sizes
-
-    (
-        h_start_index_fn,
-        h_end_index_fn,
-        w_start_index_fn,
-        w_end_index_fn,
-    ) = compute_indices_adaptive_pooling(
-        start_index, end_index, h_in, w_in, h_out, w_out
-    )
-
-    def fn(idx, loader):
-        *prefix, bh, bw = idx
-
-        h_start_index = h_start_index_fn(bh)
-        h_end_index = h_end_index_fn(bh)
-
-        w_start_index = w_start_index_fn(bw)
-        w_end_index = w_end_index_fn(bw)
-
-        result = None
-        for ih, iw in itertools.product(range(kernel_maxes[0]), range(kernel_maxes[1])):
-            val = loader(
-                prefix,
-                [ih, iw],
-                [h_start_index, w_start_index],
-                [h_end_index, w_end_index],
-            )
-            if result is None:
-                result = val
-            else:
-                result = pooling_fn(val, result)
-        return result
-
-    return fn
-
-
-def _adaptive_pooling_fn_with_idx(
-    start_index, end_index, kernel_maxes, in_sizes, out_sizes, pooling_fn
-):
-    h_in, w_in = in_sizes
-    h_out, w_out = out_sizes
-
-    (
-        h_start_index_fn,
-        h_end_index_fn,
-        w_start_index_fn,
-        w_end_index_fn,
-    ) = compute_indices_adaptive_pooling(
-        start_index, end_index, h_in, w_in, h_out, w_out
-    )
-
-    def fn(idx, loader):
-        *prefix, bh, bw = idx
-
-        h_start_index = h_start_index_fn(bh)
-        h_end_index = h_end_index_fn(bh)
-
-        w_start_index = w_start_index_fn(bw)
-        w_end_index = w_end_index_fn(bw)
-
-        maxval = None
-        maxindex = None
-        for ih, iw in itertools.product(range(kernel_maxes[0]), range(kernel_maxes[1])):
-            val = loader(
-                prefix,
-                [ih, iw],
-                [h_start_index, w_start_index],
-                [h_end_index, w_end_index],
-            )
-
-            index = ops.index_expr(
-                (h_start_index + ih) * w_in + w_start_index + iw, torch.int64
-            )
-
-            if maxindex is None:
-                maxindex = index
-            else:
-                maxindex = ops.where(ops.gt(val, maxval), index, maxindex)
-
-            if maxval is None:
-                maxval = val
-            else:
-                maxval = pooling_fn(val, maxval)
-
-        return maxindex
-
-    return fn
-
-
-fallback_adaptive_avg_pool2d = fallback_handler(
-    aten._adaptive_avg_pool2d.default, add_to_fallback_set=False
-)
-
-
-@register_lowering(aten._adaptive_avg_pool2d)
-def _adaptive_avg_pool2d(x, output_size):
-    if x.get_dtype() == torch.int64:
-        # not supported in eager
-        raise RuntimeError("'adaptive_avg_pool2d' not implemented for 'Long'")
-    assert isinstance(x, TensorBox)
-    assert len(output_size) == 2
-    x.realize_hint()
-
-    *batch, h_in, w_in = x.get_size()
-
-    h_in = V.graph.sizevars.guard_int(h_in)
-    w_in = V.graph.sizevars.guard_int(w_in)
-
-    h_out, w_out = output_size
-
-    # no-op if the same input and output
-    if h_in == h_out and w_in == w_out:
-        return clone(x)
-
-    if h_out == 0 or w_out == 0:
-        o_size = [*batch, h_out, w_out]
-        return empty(o_size, dtype=x.get_dtype(), device=x.get_device())
-    if h_in % h_out == 0 and w_in % w_out == 0:
-        kernel_size = [FloorDiv(h_in, h_out), FloorDiv(w_in, w_out)]
-        return avg_pool2d(x, kernel_size)
-
-    h_kernel_max = ceildiv((h_in + h_out - 1), h_out)
-    w_kernel_max = ceildiv((w_in + w_out - 1), w_out)
-
-    new_size = list(batch) + [h_out, w_out]
-    dtype = x.get_dtype()
-
-    window_size = h_kernel_max * w_kernel_max
-    if window_size > 25:
-        # Kernel size too big. Results in hard-to-optimize Triton code. Use fallback.
-        return fallback_adaptive_avg_pool2d(x, output_size)
-
-    def start_index(index, out_dim, inp_dim):
-        return FloorDiv((index * inp_dim), out_dim)
-
-    def end_index(index, out_dim, inp_dim):
-        return FloorDiv((index + 1) * inp_dim + out_dim - 1, out_dim)
-
-    fn_sum = _adaptive_pooling_fn(
-        start_index=start_index,
-        end_index=end_index,
-        kernel_maxes=[h_kernel_max, w_kernel_max],
-        in_sizes=[h_in, w_in],
-        out_sizes=[h_out, w_out],
-        pooling_fn=ops.add,
-    )
-
-    ones_loader = pad_adaptive_loader(ones_like(x))
-
-    def fn(idx):
-        return ops.truediv(
-            fn_sum(idx, pad_adaptive_loader(x)), fn_sum(idx, ones_loader)
-        )
-
-    rv = Pointwise.create(
-        device=x.get_device(),
-        dtype=dtype,
-        inner_fn=fn,
-        ranges=new_size,
-    )
-    # TODO: should we force these to be realized?
-    return rv
-
-
-fallback_adaptive_max_pool2d = fallback_handler(
-    aten.adaptive_max_pool2d.default, add_to_fallback_set=False
-)
-
-
-@register_lowering(aten.adaptive_max_pool2d)
-def adaptive_max_pool2d(x, output_size):
-    if x.get_dtype() == torch.int64:
-        # not supported in eager
-        raise RuntimeError("adaptive_max_pool2d not implemented for Long")
-    assert isinstance(x, TensorBox)
-    assert len(output_size) == 2
-    x.realize_hint()
-
-    *batch, h_in, w_in = x.get_size()
-
-    h_in = V.graph.sizevars.guard_int(h_in)
-    w_in = V.graph.sizevars.guard_int(w_in)
-
-    h_out, w_out = output_size
-
-    if h_out == 0 or w_out == 0:
-        o_size = [*batch, h_out, w_out]
-        return empty(o_size, dtype=x.get_dtype(), device=x.get_device()), empty(
-            o_size, dtype=torch.int64, device=x.get_device()
-        )
-
-    if h_in % h_out == 0 and w_in % w_out == 0:
-        # This is handled by a decomposition
-        raise ValueError
-
-    h_kernel_max = ceildiv((h_in + h_out - 1), h_out)
-    w_kernel_max = ceildiv((w_in + w_out - 1), w_out)
-
-    new_size = list(batch) + [h_out, w_out]
-    dtype = x.get_dtype()
-
-    window_size = h_kernel_max * w_kernel_max
-    if window_size > 25:
-        # Kernel size too big. Results in hard-to-optimize Triton code. Use fallback.
-        return fallback_adaptive_max_pool2d(x, output_size)
-
-    def start_index(index, out_dim, inp_dim):
-        return FloorDiv((index * inp_dim), out_dim)
-
-    def end_index(index, out_dim, inp_dim):
-        return FloorDiv((index + 1) * inp_dim + out_dim - 1, out_dim)
-
-    inner_func_max_val = _adaptive_pooling_fn(
-        start_index=start_index,
-        end_index=end_index,
-        kernel_maxes=[h_kernel_max, w_kernel_max],
-        in_sizes=[h_in, w_in],
-        out_sizes=[h_out, w_out],
-        pooling_fn=ops.maximum,
-    )
-
-    inner_func_max_idx = _adaptive_pooling_fn_with_idx(
-        start_index=start_index,
-        end_index=end_index,
-        kernel_maxes=[h_kernel_max, w_kernel_max],
-        in_sizes=[h_in, w_in],
-        out_sizes=[h_out, w_out],
-        pooling_fn=ops.maximum,
-    )
-
-    def inner_fn_max_val(idx):
-        return inner_func_max_val(idx, pad_adaptive_loader(x, float("-inf")))
-
-    def inner_fn_max_idx(idx):
-        return inner_func_max_idx(idx, pad_adaptive_loader(x, float("-inf")))
-
-    rv = Pointwise.create(
-        device=x.get_device(),
-        dtype=dtype,
-        inner_fn=inner_fn_max_val,
-        ranges=new_size,
-    )
-    ri = Pointwise.create(
-        device=x.get_device(),
-        dtype=torch.int64,
-        inner_fn=inner_fn_max_idx,
-        ranges=new_size,
-    )
-    return rv, ri
-
-
-def _fractional_pooling_offsets(samples, in_sz, out_sz, kernel_sz, dim, ndims):
-    out_sz = out_sz[dim]
-    in_sz = in_sz[dim]
-    kernel_sz = kernel_sz[dim]
-    samples_loader = samples.make_loader()
-
-    def load(prefix, i):
-        # Handle indexing for samples tensor correctly for different input dimensions
-        # samples tensor always has shape (N, C, 2) for fractional_max_pool2d where:
-        # - N=1 for 3D inputs (C,H,W), N=batch_size for 4D inputs (N,C,H,W)
-        # - C=num_channels
-        # - 2 for the two spatial dimensions (height, width)
-        samples_shape = samples.get_size()
-
-        if len(samples_shape) == 3:  # Expected: (N, C, 2)
-            if len(prefix) == 1:
-                # 3D input case: prefix=(channel,), samples=(1, C, 2)
-                # Access: samples[0, channel, dim]
-                sample = samples_loader([0, prefix[0], ndims - 1 - dim])
-            elif len(prefix) >= 2:
-                # 4D+ input case: prefix=(batch, channel, ...), samples=(batch, C, 2)
-                # Access: samples[batch, channel, dim]
-                sample = samples_loader([prefix[0], prefix[1], ndims - 1 - dim])
-            else:
-                # Edge case - shouldn't happen for valid fractional pooling
-                sample = samples_loader([0, 0, ndims - 1 - dim])
-        else:
-            # Fallback for unexpected tensor shapes
-            sample = samples_loader([*prefix, ndims - 1 - dim])
-        i_expr = ops.index_expr(i, samples.get_dtype())
-        diff = ops.index_expr(in_sz - kernel_sz, torch.int64)
-        out_sz_expr = ops.index_expr(out_sz - 1, torch.int64)
-        alpha = ops.truediv(
-            ops.to_dtype(diff, torch.float64), ops.to_dtype(out_sz_expr, torch.float64)
-        )
-        alpha = ops.where(ops.eq(out_sz_expr, 0), 0, alpha)
-        seq_i = ops.trunc((i_expr + sample) * alpha) - ops.trunc(sample * alpha)
-        seq_i = ops.to_dtype(seq_i, torch.int64)
-        mask = ops.lt(i_expr, out_sz_expr)
-        return ops.indirect_indexing(ops.where(mask, seq_i, diff), sympy.sympify(in_sz))
-
-    return load
-
-
-@register_lowering(aten.fractional_max_pool2d)
-def fractional_max_pool2d(x, kernel_size, output_size, random_samples):
-    return _fractional_max_pool(x, kernel_size, output_size, random_samples, n_dim=2)
-
-
-@register_lowering(aten.fractional_max_pool3d)
-def fractional_max_pool3d(x, kernel_size, output_size, random_samples):
-    return _fractional_max_pool(x, kernel_size, output_size, random_samples, n_dim=3)
-
-
-def _fractional_max_pool(x, kernel_size, output_size, random_samples, n_dim):
-    x.realize_hint()
-    batch, inp_dhw = x.shape[:-n_dim], x.shape[-n_dim:]
-
-    with config.patch(unroll_reductions_threshold=25):
-        dhw_index_fn = [
-            _fractional_pooling_offsets(
-                samples=random_samples,
-                in_sz=inp_dhw,
-                out_sz=output_size,
-                kernel_sz=kernel_size,
-                ndims=n_dim,
-                dim=d,
-            )
-            for d in range(n_dim)
-        ]
-
-        x_loader = x.make_loader()
-
-        def fn_inner(idx, reduction_idx):
-            prefix = idx[:-n_dim]
-            return x_loader([*prefix, *increments_to_index(idx, reduction_idx)])
-
-        def increments_to_index(idx, reduction_idx):
-            prefix = idx[:-n_dim]
-            bdhw = idx[-n_dim:]
-            return [
-                dhw_index_fn[d](prefix, bdhw[d]) + reduction_idx[d]
-                for d in range(n_dim)
-            ]
-
-        new_size = list(batch) + list(output_size)
-        dtype = x.get_dtype()
-        result = Reduction.create(
-            reduction_type="max",
-            input_node=x,
-            device=x.get_device(),
-            dst_dtype=dtype,
-            src_dtype=dtype,
-            inner_fn=fn_inner,
-            ranges=new_size,
-            reduction_ranges=kernel_size,
-        )
-        offsets = Reduction.create(
-            reduction_type="argmax",
-            input_node=x,
-            device=x.get_device(),
-            dst_dtype=torch.int64,
-            src_dtype=dtype,
-            inner_fn=fn_inner,
-            ranges=new_size,
-            reduction_ranges=kernel_size,
-        )
-        assert isinstance(result, TensorBox), result
-        if isinstance(result.data.data, Reduction):  # type: ignore[attr-defined]
-            # Only realize if reduction isn't unrolled
-            result.realize()
-        assert isinstance(offsets, TensorBox), offsets
-        if isinstance(offsets.data.data, Reduction):  # type: ignore[attr-defined]
-            # Only realize if reduction isn't unrolled
-            offsets.realize()
-
-        indices = _pool_offsets_to_indices(
-            offsets, kernel_size, x.shape, increments_to_index
-        )
-        return result, indices
-
-
-@register_lowering(aten.upsample_nearest2d_backward.default)
-def upsample_nearest2d_backward(
-    x, output_size=None, input_size=None, scales_h=None, scales_w=None
-):
-    x.realize_hint()
-
-    *_batch, inp_h, inp_w = x.get_size()
-    inp_h = V.graph.sizevars.guard_int(inp_h)
-    inp_w = V.graph.sizevars.guard_int(inp_w)
-
-    # pyrefly: ignore [not-iterable]
-    *_batch, out_h, out_w = input_size
-
-    if inp_h % out_h == 0 and inp_w % out_w == 0:
-        return avg_pool2d(
-            x, [FloorDiv(inp_h, out_h), FloorDiv(inp_w, out_w)], divisor_override=1
-        )
-
-    h_kernel_max = ceildiv(inp_h, out_h)
-    w_kernel_max = ceildiv(inp_w, out_w)
-
-    def start_index(index, out_dim, inp_dim):
-        return CeilDiv(index * inp_dim, sympy.sympify(out_dim))
-
-    def end_index(index, out_dim, inp_dim):
-        return start_index((index + 1), out_dim, inp_dim)
-
-    fn_sum = _adaptive_pooling_fn(
-        start_index=start_index,
-        end_index=end_index,
-        kernel_maxes=[h_kernel_max, w_kernel_max],
-        in_sizes=[inp_h, inp_w],
-        out_sizes=[out_h, out_w],
-        pooling_fn=ops.add,
-    )
-
-    def fn(idx):
-        return fn_sum(idx, pad_adaptive_loader(x))
-
-    rv = Pointwise.create(
-        device=x.get_device(),
-        dtype=x.get_dtype(),
-        inner_fn=fn,
-        # pyrefly: ignore [bad-argument-type, no-matching-overload]
-        ranges=list(input_size),
-    )
-
-    return rv
-
-
-@register_lowering(aten.avg_pool2d, type_promotion_kind=None)
-def avg_pool2d(
-    x,
-    kernel_size,
-    stride=(),
-    padding=0,
-    ceil_mode=False,
-    count_include_pad=True,
-    divisor_override=None,
-):
-    return _avg_poolnd(
-        x,
-        kernel_size,
-        stride,
-        padding,
-        ceil_mode,
-        count_include_pad,
-        divisor_override,
-        dim=2,
-    )
-
-
-@register_lowering(aten.avg_pool3d, type_promotion_kind=None)
-def avg_pool3d(
-    x,
-    kernel_size,
-    stride=(),
-    padding=0,
-    ceil_mode=False,
-    count_include_pad=True,
-    divisor_override=None,
-):
-    return _avg_poolnd(
-        x,
-        kernel_size,
-        stride,
-        padding,
-        ceil_mode,
-        count_include_pad,
-        divisor_override,
-        dim=3,
-    )
-
-
-fallbacks_avg_poolnd = [
-    fallback_handler(aten.avg_pool1d.default, add_to_fallback_set=False),
-    fallback_handler(aten.avg_pool2d.default, add_to_fallback_set=False),
-    fallback_handler(aten.avg_pool3d.default, add_to_fallback_set=False),
-]
-
-
-def _avg_poolnd(
-    x,
-    kernel_size,
-    stride,
-    padding,
-    ceil_mode,
-    count_include_pad,
-    divisor_override,
-    dim,
-):
-    if not stride:
-        stride = kernel_size
-    if not padding:
-        padding = [0] * dim
-    kernel_size = pad_listlike(kernel_size, dim)
-    stride = pad_listlike(stride, dim)
-    padding = pad_listlike(padding, dim)
-
-    assert isinstance(x, TensorBox)
-    assert len(kernel_size) == dim
-    assert len(stride) == dim
-    assert len(padding) == dim
-    assert len(x.get_size()) in (dim + 1, dim + 2)
-
-    x.realize_hint()
-    batch = x.get_size()[:-dim]
-    h = x.get_size()[-dim:]
-
-    h_out, ceil_modes = zip(
-        *[
-            pooling_size(h[i], i, kernel_size, stride, padding, ceil_mode)
-            for i in range(dim)
-        ]
-    )
-
-    if any(padding) or any(ceil_modes):
-        x_loader = constant_boundary_condition(x, 0.0, dim=dim)
-        had_padding = True
-    else:
-        x_loader = x.make_loader()
-        had_padding = False
-
-    new_size = list(batch) + list(h_out)
-    dtype = x.get_dtype()
-    # compute in higher-precision until scaling
-    output_dtype = get_promoted_dtype(
-        x,
-        type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.DEFAULT,
-        return_compute_dtype=True,
-    )
-
-    def fn_inner(idx, reduction_idx):
-        prefix = idx[:-dim]
-        bh = idx[-dim:]
-        ih = reduction_idx
-        ih = [bh[i] * stride[i] + ih[i] - padding[i] for i in range(dim)]
-        return x_loader([*prefix, *ih])
-
-    window_size = functools.reduce(operator.mul, kernel_size)
-
-    if window_size > 25 and any(
-        V.graph.sizevars.statically_known_true(sympy.Ne(k, s))
-        for k, s in zip(kernel_size, stride)
-    ):
-        fallback = fallbacks_avg_poolnd[dim - 1]
-        return fallback(
-            x,
-            kernel_size,
-            stride,
-            padding,
-            ceil_mode,
-            count_include_pad,
-            divisor_override,
-        )
-
-    # TODO: remove this when #100331 is merged. We only do this
-    # for window_size <=25 to avoid performance regressions compared
-    # to the previous algorithm which unrolled manually for <=25
-    context = (
-        config.patch(unroll_reductions_threshold=25)
-        if window_size <= 25
-        else contextlib.nullcontext()
-    )
-
-    device = x.get_device()
-    assert device is not None
-
-    with context:
-        rv = Reduction.create(
-            reduction_type="sum",
-            input_node=x,
-            device=device,
-            dst_dtype=output_dtype,
-            src_dtype=dtype,
-            inner_fn=fn_inner,
-            ranges=new_size,
-            reduction_ranges=kernel_size,
-        )
-    if hasattr(rv.data, "data") and isinstance(rv.data.data, Reduction):
-        # Only realize if reduction isn't unrolled
-        rv.realize()
-
-    if not had_padding or divisor_override:
-        divisor = divisor_override if divisor_override else window_size
-        result = div_prim(rv, divisor)
-    else:
-
-        def fn_count(idx):
-            bh = idx[-dim:]
-
-            divide_factors = []
-            for i in range(dim):
-                hstart = bh[i] * stride[i] - padding[i]
-                hend = sympy.Min(hstart + kernel_size[i], h[i] + padding[i])
-                if not count_include_pad:
-                    hstart = sympy.Max(hstart, 0)
-                    hend = sympy.Min(hend, h[i])
-                factor = ops.index_expr(hend - hstart, torch.int32)
-                divide_factors.append(factor)
-            return functools.reduce(ops.mul, divide_factors)
-
-        divide_factor = Pointwise.create(
-            device=x.get_device(),
-            dtype=dtype,
-            inner_fn=fn_count,
-            ranges=new_size,
-        )
-        result = div_prim(rv, divide_factor)
-
-    return to_dtype(result, dtype)
-
-
-fallback_avg_pool2d_backward = fallback_handler(
-    aten.avg_pool2d_backward.default, add_to_fallback_set=False
-)
-
-
-@register_lowering(aten.avg_pool2d_backward, type_promotion_kind=None)
-def avg_pool2d_backward(
-    grad_output,
-    x,
-    kernel_size,
-    stride,
-    padding,
-    ceil_mode,
-    count_include_pad,
-    divisor_override=None,
-):
-    assert divisor_override is None or divisor_override != 0, "divisor must be not zero"
-    if not stride:
-        stride = kernel_size
-    if not padding:
-        padding = [0, 0]
-
-    assert isinstance(grad_output, TensorBox)
-    assert isinstance(x, TensorBox)
-    assert len(kernel_size) == 2
-    assert len(stride) == 2
-    assert len(padding) == 2
-    assert len(x.get_size()) in (3, 4)
-
-    grad_output.realize_hint()  # we will read this many times, so make sure it is computed
-
-    *_, height, width = x.get_size()
-
-    _h_out, ceil_mode1 = pooling_size(
-        height, 0, kernel_size, stride, padding, ceil_mode
-    )
-    _w_out, ceil_mode2 = pooling_size(width, 1, kernel_size, stride, padding, ceil_mode)
-
-    grad_loader = grad_output.make_loader()
-
-    had_padding = padding[0] or padding[1] or ceil_mode1 or ceil_mode2
-
-    *_, pooled_height, pooled_width = grad_output.get_size()
-    new_size = list(x.get_size())
-    dtype = x.get_dtype()
-
-    h_window_size = max(
-        max(FloorDiv(h, stride[0]) - max(0, FloorDiv(h - kernel_size[0], stride[0])), 1)
-        for h in range(kernel_size[0] * 2)
-    )
-    w_window_size = max(
-        max(FloorDiv(w, stride[1]) - max(0, FloorDiv(w - kernel_size[1], stride[1])), 1)
-        for w in range(kernel_size[1] * 2)
-    )
-
-    window_size = h_window_size * w_window_size
-    if window_size > 25:
-        # Kernel size too big. Results in hard-to-optimize Triton code. Use fallback.
-        return fallback_avg_pool2d_backward(
-            grad_output,
-            x,
-            kernel_size,
-            stride,
-            padding,
-            ceil_mode,
-            count_include_pad,
-            divisor_override,
-        )
-
-    def compute_pool_size_without_padding(ph, pw):
-        """
-        This computes the scaling factor that we will divide an element
-        by when `count_include_pad=False`
-        """
-        stride_h = ops.constant(stride[0], torch.int32)
-        stride_w = ops.constant(stride[1], torch.int32)
-        pad_h = ops.constant(padding[0], torch.int32)
-        pad_w = ops.constant(padding[1], torch.int32)
-        kernel_h = ops.constant(kernel_size[0], torch.int32)
-        kernel_w = ops.constant(kernel_size[1], torch.int32)
-        hstart = ops.sub(ops.mul(ph, stride_h), pad_h)
-        wstart = ops.sub(ops.mul(pw, stride_w), pad_w)
-        hend = ops.minimum(
-            ops.add(hstart, kernel_h),
-            ops.add(ops.index_expr(height, torch.int32), pad_h),
-        )
-        wend = ops.minimum(
-            ops.add(wstart, kernel_w),
-            ops.add(ops.index_expr(width, torch.int32), pad_w),
-        )
-        hstart = ops.maximum(hstart, ops.constant(0, torch.int32))
-        wstart = ops.maximum(wstart, ops.constant(0, torch.int32))
-        hend = ops.minimum(hend, ops.index_expr(height, torch.int32))
-        wend = ops.minimum(wend, ops.index_expr(width, torch.int32))
-        divide_factor = ops.mul(ops.sub(hend, hstart), ops.sub(wend, wstart))
-        return divide_factor
-
-    def fn(idx):
-        *prefix, h, w = idx
-        h = h + padding[0]
-        w = w + padding[1]
-        phstart = ops.index_expr(
-            FloorDiv(h - kernel_size[0] + stride[0], stride[0]), torch.int32
-        )
-        pwstart = ops.index_expr(
-            FloorDiv(w - kernel_size[1] + stride[1], stride[1]), torch.int32
-        )
-        phend = ops.index_expr(FloorDiv(h, stride[0]) + 1, torch.int32)
-        pwend = ops.index_expr(FloorDiv(w, stride[1]) + 1, torch.int32)
-
-        phstart = ops.maximum(phstart, ops.constant(0, torch.int32))
-        pwstart = ops.maximum(pwstart, ops.constant(0, torch.int32))
-        phend = ops.minimum(phend, ops.index_expr(pooled_height, torch.int32))
-        pwend = ops.minimum(pwend, ops.index_expr(pooled_width, torch.int32))
-
-        gradient = None
-        for ph_ in range(h_window_size):
-            for pw_ in range(w_window_size):
-                ph = ops.add(phstart, ops.constant(ph_, torch.int32))
-                pw = ops.add(pwstart, ops.constant(pw_, torch.int32))
-
-                if divisor_override is not None:
-                    scale = divisor_override
-                elif count_include_pad or not had_padding:
-                    scale = kernel_size[0] * kernel_size[1]
-                else:
-                    scale = compute_pool_size_without_padding(ph, pw)
-
-                part = ops.truediv(
-                    grad_loader(
-                        [
-                            *prefix,
-                            ops.indirect_indexing(
-                                ops.minimum(
-                                    ph, ops.sub(phend, ops.constant(1, torch.int32))
-                                ),
-                                pooled_height,
-                                check=False,
-                            ),
-                            ops.indirect_indexing(
-                                ops.minimum(
-                                    pw, ops.sub(pwend, ops.constant(1, torch.int32))
-                                ),
-                                pooled_width,
-                                check=False,
-                            ),
-                        ]
-                    ),
-                    scale,
-                )
-
-                mask = ops.and_(
-                    ops.lt(ph, phend),
-                    ops.lt(pw, pwend),
-                )
-                if gradient is None:
-                    gradient = ops.where(mask, part, ops.constant(0.0, torch.float32))
-                else:
-                    gradient = ops.where(mask, ops.add(gradient, part), gradient)
-        assert gradient is not None
-        return gradient
-
-    rv = Pointwise.create(
-        device=grad_output.get_device(),
-        dtype=dtype,
-        inner_fn=fn,
-        ranges=new_size,
-    )
-    return rv
-
-
-fallback_avg_pool3d_backward = fallback_handler(
-    aten.avg_pool3d_backward.default, add_to_fallback_set=False
-)
-
-
-@register_lowering(aten.avg_pool3d_backward, type_promotion_kind=None)
-def avg_pool3d_backward(
-    grad_output,
-    x,
-    kernel_size,
-    stride,
-    padding,
-    ceil_mode,
-    count_include_pad,
-    divisor_override=None,
-):
-    assert divisor_override is None or divisor_override != 0, "divisor must be not zero"
-    if not stride:
-        stride = kernel_size
-    if not padding:
-        padding = [0, 0, 0]
-
-    assert isinstance(grad_output, TensorBox)
-    assert isinstance(x, TensorBox)
-    assert len(kernel_size) == 3
-    assert len(stride) == 3
-    assert len(padding) == 3
-    assert len(x.get_size()) in (4, 5)
-
-    grad_output.realize_hint()
-
-    *_batch, depth, height, width = x.get_size()
-
-    _d_out, ceil_mode_d = pooling_size(
-        depth, 0, kernel_size, stride, padding, ceil_mode
-    )
-    _h_out, ceil_mode_h = pooling_size(
-        height, 1, kernel_size, stride, padding, ceil_mode
-    )
-    _w_out, ceil_mode_w = pooling_size(
-        width, 2, kernel_size, stride, padding, ceil_mode
-    )
-
-    grad_loader = grad_output.make_loader()
-    had_padding = any(padding) or ceil_mode_d or ceil_mode_h or ceil_mode_w
-
-    *_, pooled_depth, pooled_height, pooled_width = grad_output.get_size()
-    new_size = list(x.get_size())
-    dtype = x.get_dtype()
-
-    d_window_size, h_window_size, w_window_size = (
-        max(
-            max(d // stride[i] - max(0, (d - kernel_size[i]) // stride[i]), 1)
-            for d in range(kernel_size[i] * 2)
-        )
-        for i in range(3)
-    )
-
-    window_size = d_window_size * h_window_size * w_window_size
-    if window_size > 125:
-        # Kernel size too big. Results in hard-to-optimize Triton code.
-        return fallback_avg_pool3d_backward(
-            grad_output,
-            x,
-            kernel_size,
-            stride,
-            padding,
-            ceil_mode,
-            count_include_pad,
-            divisor_override,
-        )
-
-    def compute_pool_size_without_padding(pd, ph, pw):
-        stride_d, stride_h, stride_w = (ops.constant(s, torch.int32) for s in stride)
-        pad_d, pad_h, pad_w = (ops.constant(p, torch.int32) for p in padding)
-        kernel_d, kernel_h, kernel_w = (
-            ops.constant(k, torch.int32) for k in kernel_size
-        )
-
-        dstart, hstart, wstart = (
-            ops.sub(ops.mul(p, s), pad)
-            for p, s, pad in zip(
-                [pd, ph, pw], [stride_d, stride_h, stride_w], [pad_d, pad_h, pad_w]
-            )
-        )
-        dend, hend, wend = (
-            ops.minimum(
-                ops.add(start, k), ops.add(ops.index_expr(dim, torch.int32), pad)
-            )
-            for start, k, dim, pad in zip(
-                [dstart, hstart, wstart],
-                [kernel_d, kernel_h, kernel_w],
-                [depth, height, width],
-                [pad_d, pad_h, pad_w],
-            )
-        )
-        dstart, hstart, wstart = (
-            ops.maximum(start, ops.constant(0, torch.int32))
-            for start in [dstart, hstart, wstart]
-        )
-        dend, hend, wend = (
-            ops.minimum(end, ops.index_expr(dim, torch.int32))
-            for end, dim in zip([dend, hend, wend], [depth, height, width])
-        )
-        divide_factor = ops.mul(
-            ops.mul(ops.sub(dend, dstart), ops.sub(hend, hstart)), ops.sub(wend, wstart)
-        )
-        return divide_factor
-
-    def fn(idx):
-        *prefix, d, h, w = idx
-        d, h, w = (v + pad for v, pad in zip([d, h, w], padding))
-
-        pdstart, phstart, pwstart = (
-            ops.index_expr(FloorDiv(v - k + s, s), torch.int32)
-            for v, k, s in zip([d, h, w], kernel_size, stride)
-        )
-
-        pdend, phend, pwend = (
-            ops.index_expr(FloorDiv(v, s) + 1, torch.int32)
-            for v, s in zip([d, h, w], stride)
-        )
-
-        pdstart, phstart, pwstart = (
-            ops.maximum(pstart, ops.constant(0, torch.int32))
-            for pstart in [pdstart, phstart, pwstart]
-        )
-        pdend, phend, pwend = (
-            ops.minimum(pend, ops.index_expr(pooled_dim, torch.int32))
-            for pend, pooled_dim in zip(
-                [pdend, phend, pwend], [pooled_depth, pooled_height, pooled_width]
-            )
-        )
-
-        gradient = None
-        # Iterate over the 3D region to accumulate gradients
-        for pd_ in range(d_window_size):
-            for ph_ in range(h_window_size):
-                for pw_ in range(w_window_size):
-                    pd, ph, pw = (
-                        ops.add(pstart, ops.constant(p_, torch.int32))
-                        for pstart, p_ in zip(
-                            [pdstart, phstart, pwstart], [pd_, ph_, pw_]
-                        )
-                    )
-
-                    if divisor_override is not None:
-                        scale = divisor_override
-                    elif count_include_pad or not had_padding:
-                        scale = kernel_size[0] * kernel_size[1] * kernel_size[2]
-                    else:
-                        scale = compute_pool_size_without_padding(pd, ph, pw)
-
-                    part = ops.truediv(
-                        grad_loader(
-                            [
-                                *prefix,
-                                ops.indirect_indexing(
-                                    ops.minimum(
-                                        pd, ops.sub(pdend, ops.constant(1, torch.int32))
-                                    ),
-                                    pooled_depth,
-                                    check=False,
-                                ),
-                                ops.indirect_indexing(
-                                    ops.minimum(
-                                        ph, ops.sub(phend, ops.constant(1, torch.int32))
-                                    ),
-                                    pooled_height,
-                                    check=False,
-                                ),
-                                ops.indirect_indexing(
-                                    ops.minimum(
-                                        pw, ops.sub(pwend, ops.constant(1, torch.int32))
-                                    ),
-                                    pooled_width,
-                                    check=False,
-                                ),
-                            ]
-                        ),
-                        scale,
-                    )
-
-                    mask = ops.and_(
-                        ops.and_(ops.lt(pd, pdend), ops.lt(ph, phend)),
-                        ops.lt(pw, pwend),
-                    )
-                    if gradient is None:
-                        gradient = ops.where(
-                            mask, part, ops.constant(0.0, torch.float32)
-                        )
-                    else:
-                        gradient = ops.where(mask, ops.add(gradient, part), gradient)
-        assert gradient is not None
-        return gradient
-
-    rv = Pointwise.create(
-        device=grad_output.get_device(),
-        dtype=dtype,
-        inner_fn=fn,
-        ranges=new_size,
-    )
-    return rv
-
-
-def _validate_reduction_axis(x, axis):
-    size = x.get_size()
-    if isinstance(axis, int):
-        axis = [axis]
-    elif not axis:
-        axis = range(len(size))
-    if len(size) == 0:
-        assert tuple(axis) in [(), (0,), (-1,)], f"invalid axis: {axis}"
-        return []
-    axis = list(axis)
-    for i in range(len(axis)):
-        if axis[i] < 0:
-            axis[i] += len(size) if len(size) else 1
-        assert 0 <= axis[i] < len(size) or (len(size) == 0 and axis[i] == 0)
-    assert len(OrderedSet(axis)) == len(axis), "reduction axis not unique"
-    return axis
-
-
-def _make_reduction_inner(
-    x, *, axis, keepdims, dtype, override_return_dtype, reduction_type=None
-):
-    if dtype is not None:
-        x = to_dtype(x, dtype)
-    size = x.get_size()
-    axis = OrderedSet[int](_validate_reduction_axis(x, axis))
-
-    kept_sizes = []
-    kept_idx = []
-    reduced_sizes = []
-    reduced_idx = []
-    for i in range(len(size)):
-        if i in axis:
-            reduced_idx.append(i)
-            reduced_sizes.append(size[i])
-        else:
-            kept_idx.append(i)
-            kept_sizes.append(size[i])
-
-    # For argmax/argmin compute logical indices when the tensor has non-contiguous layout.
-    should_compute_logical_index = False
-    if (
-        reduction_type in ("argmax", "argmin")
-        and len(reduced_sizes) > 1
-        and is_triton(x)
-    ):
-        if isinstance(x.data, PermuteView):
-            should_compute_logical_index = True
-        elif isinstance(x.data, ir.ReinterpretView) or (
-            isinstance(x.data, ir.StorageBox) and isinstance(x.data.data, ir.Buffer)
-        ):
-            layout = x.get_layout()
-            should_compute_logical_index = (
-                layout.is_transposed() or not layout.is_contiguous()
-            )
-
-    def loader(index, reduction_index):
-        assert len(reduction_index) == len(reduced_idx)
-        if keepdims:
-            assert len(index) == len(size)
-            index = [index[i] for i in kept_idx]
-        assert len(index) == len(kept_idx)
-        new_index = [None] * (len(index) + len(reduction_index))
-        for idx, var in itertools.chain(
-            zip(kept_idx, index), zip(reduced_idx, reduction_index)
-        ):
-            new_index[idx] = var
-        value = inner_loader(new_index)
-
-        # For argmax/argmin, return tuple with logical linear index if needed
-        if should_compute_logical_index:
-            rindex = [sympy.expand(i) for i in reduction_index]
-
-            # Compute linear index in row-major order
-            # For reduction_ranges = [4, 6]: linear_index = r0 * 6 + r1
-            linear_idx = rindex[0]
-            for i in range(1, len(rindex)):
-                linear_idx = linear_idx * reduced_sizes[i] + rindex[i]
-
-            return (value, ops.index_expr(linear_idx, torch.int64))
-
-        return value
-
-    if keepdims:
-        new_size = list(size)
-        for i in reduced_idx:
-            new_size[i] = sympy.S.One
-    else:
-        new_size = kept_sizes
-
-    inner_loader = x.make_loader()
-    return dict(
-        device=x.get_device(),
-        dst_dtype=override_return_dtype or x.get_dtype(),
-        src_dtype=x.get_dtype(),
-        inner_fn=loader,
-        ranges=new_size,
-        reduction_ranges=reduced_sizes,
-    )
-
-
-def make_reduction(reduction_type: ReductionType, override_return_dtype=None):
-    def inner(x, axis=None, keepdims=False, *, dtype=None):
-        # For argmax/argmin on boolean tensors, cast to int32 first to ensure
-        # correct comparison in Triton. See https://github.com/pytorch/pytorch/issues/174069
-        # Only apply on Triton backend; MPS handles bool comparisons natively.
-        if (
-            reduction_type in ("argmax", "argmin")
-            and x.get_dtype() == torch.bool
-            and is_triton(x)
-        ):
-            x = to_dtype(x, torch.int32)
-        kwargs = _make_reduction_inner(
-            x,
-            axis=axis,
-            keepdims=keepdims,
-            dtype=dtype,
-            override_return_dtype=override_return_dtype,
-            reduction_type=reduction_type,
-        )
-        result = Reduction.create(reduction_type=reduction_type, input_node=x, **kwargs)
-        if isinstance(
-            result.data.data,  # type: ignore[attr-defined, attr-type, union-attr]
-            Reduction,
-        ):  # Only realize if reduction isn't unrolled
-            result.realize()
-        return result
-
-    return inner
-
-
-def _make_scan_inner(x, *, axis, dtype):
-    if dtype is not None:
-        x = to_dtype(x, dtype)
-    axis = _validate_dim(x, axis)
-
-    return dict(
-        device=x.get_device(),
-        dtypes=(x.get_dtype(),),
-        inner_fns=(x.make_loader(),),
-        size=x.get_size(),
-        axis=axis,
-    )
-
-
-@register_lowering(aten.mean)
-def mean(x, axis=None, keepdim=False, *, dtype=None):
-    if dtype is not None:
-        x = to_dtype(x, dtype)
-    size = x.get_size()
-    axis = _validate_reduction_axis(x, axis)
-    # compute in higher-precision until end of mean lowering
-    output_dtype = x.get_dtype()
-    if output_dtype in (torch.float16, torch.bfloat16):
-        x = to_dtype(x, torch.float)
-    sum_result = sum_(x, axis, keepdim)
-    denom = sympy_product(size[i] for i in axis)
-    denom = ir.IndexingConstant(index=denom, dtype=x.get_dtype(), device=x.get_device())
-    denom = ExpandView.create(denom, list(sum_result.get_size()))
-    return to_dtype(div(sum_result, denom), output_dtype)
-
-
-def var_mean_sum_(x, axis, correction, keepdim, return_mean):
-    if correction is None:
-        correction = 1
-
-    size = x.get_size()
-    axis = _validate_reduction_axis(x, axis)
-    x_mean = mean(x, axis, keepdim=True)
-    if return_mean:
-        x_mean.realize()
-
-    diffs = square(sub(x, x_mean))
-    sum_result = sum_(diffs, axis, keepdim)
-
-    denom = sympy_product(size[i] for i in axis)
-    if correction:
-        denom = sympy.Max(denom - correction, 0)
-    denom = ir.IndexingConstant(index=denom, dtype=x.get_dtype(), device=x.get_device())
-    denom = ExpandView.create(denom, list(sum_result.get_size()))
-    x_var = div(sum_result, denom)
-    if not return_mean:
-        return (x_var,)
-
-    x_mean = x_mean if keepdim else squeeze(x_mean, axis)
-    return x_var, x_mean
-
-
-def use_two_step_variance(x, axis, keepdim):
-    # two-step algorithm can get better performance in small reductions size
-    # while it can accumulate more numerical error than Welford algorithm.
-    axis = _validate_reduction_axis(x, axis)
-    kwargs = _make_reduction_inner(
-        x, axis=axis, keepdims=keepdim, dtype=None, override_return_dtype=None
-    )
-
-    ranges = kwargs["ranges"]
-    reduction_numel = sympy_product(kwargs["reduction_ranges"])
-    device = x.get_device()
-    if not (device and device.type == "cpu"):
-        threshold = config.unroll_reductions_threshold
-    else:
-        # 1024 is a default value to pass all the UTs about accuracy.
-        # A larger threshold can still get performance benefits.
-        threshold = config.cpp.use_two_step_variance_threshold
-    return (
-        isinstance(reduction_numel, sympy.Integer)
-        and int(reduction_numel) <= threshold
-        and sympy_product(ranges) != 1
-    )
-
-
-def var_mean_welford_(x, axis, *, correction, keepdim, return_mean):
-    if correction is None:
-        correction = 1
-
-    kwargs = _make_reduction_inner(
-        x, axis=axis, keepdims=keepdim, dtype=None, override_return_dtype=None
-    )
-    loader = kwargs.pop("inner_fn")
-    kwargs.pop("dst_dtype")
-    kwargs.pop("src_dtype")
-
-    mean, m2, _ = ir.WelfordReduction.create(
-        inner_fns=(loader,),
-        reduction_type="welford_reduce",
-        dtype=x.get_dtype(),
-        **kwargs,
-    )
-    m2.realize()
-
-    dtype = x.get_dtype()
-    size = x.get_size()
-    axis = _validate_reduction_axis(x, axis)
-    rnumel = sympy_product(size[i] for i in axis)
-
-    def get_constant_or_index_expr(x, dtype):
-        if isinstance(x, sympy.Expr) and not x.is_number:
-            return ops.to_dtype(ops.index_expr(x, torch.int64), dtype)
-        return ops.constant(x, dtype)
-
-    def scale_fn(data):
-        c = get_constant_or_index_expr(correction, dtype)
-        N = get_constant_or_index_expr(rnumel, dtype)
-        zero = ops.constant(0, dtype)
-        return data / ops.maximum(zero, N - c)
-
-    var = make_pointwise(scale_fn)(m2)
-
-    if return_mean:
-        mean.realize()
-        return var, mean
-    return (var,)
-
-
-def var_mean_helper_(x, *, axis, correction, keepdim, return_mean):
-    out_dtype = x.get_dtype()
-    compute_dtype = get_computation_dtype(out_dtype)
-    x = to_dtype(x, compute_dtype, copy=False)
-    kwargs = dict(
-        x=x,
-        axis=axis,
-        correction=correction,
-        keepdim=keepdim,
-        return_mean=return_mean,
-    )
-    output = (
-        var_mean_sum_(**kwargs)
-        if (
-            config.mtia.disable_welford_reduction
-            or use_two_step_variance(x, axis=axis, keepdim=keepdim)
-        )
-        else var_mean_welford_(**kwargs)
-    )
-    output = tuple(to_dtype(x, out_dtype, copy=False) for x in output)
-    return output[0] if not return_mean else output
-
-
-@register_lowering([aten.var, prims.var])
-def var_(x, axis=None, *, correction=None, keepdim=False):
-    return var_mean_helper_(
-        x, axis=axis, correction=correction, keepdim=keepdim, return_mean=False
-    )
-
-
-@register_lowering(aten.var_mean)
-def var_mean(x, axis=None, *, correction=None, keepdim=False):
-    return var_mean_helper_(
-        x, axis=axis, correction=correction, keepdim=keepdim, return_mean=True
-    )
-
-
-def pow_recursive(x, y, dtype):
-    if y < 0:
-        return pow_recursive(ops.reciprocal(x), -y, dtype)
-    if y == 0:
-        return ops.constant(1, dtype)
-    if y == 1:
-        return x
-
-    result = pow_recursive(x, y // 2, dtype)
-    result = ops.mul(result, result)
-    if (y % 2) == 1:
-        result = ops.mul(result, x)
-    return result
-
-
-@make_pointwise
-def pow_native(a, b):
-    return ops.pow(a, b)
-
-
-fallback_pow_tensor_tensor = fallback_handler(
-    aten.pow.Tensor_Tensor, add_to_fallback_set=False
-)
-fallback_pow_scalar = fallback_handler(aten.pow.Scalar, add_to_fallback_set=False)
-fallback_pow_tensor_scalar = fallback_handler(
-    aten.pow.Tensor_Scalar, add_to_fallback_set=False
-)
-
-
-@register_lowering(aten.pow, broadcast=True)
-def pow(a, b):
-    if isinstance(b, float) and b.is_integer():
-        return pow(a, int(b))
-    elif isinstance(b, float) and b == 0.5:
-        return sqrt(a)
-    elif isinstance(b, int) and b == 1:
-        return clone(a)
-
-    # Type promotion ensures all tensor arguments have the same type
-    dtype = next(x.get_dtype() for x in (a, b) if isinstance(x, ir.TensorBox))
-    is_integer_pow = is_integer_dtype(dtype)
-
-    # Optimize away small fixed powers, or for integers avoid falling back to ATen
-    embed_exponent = isinstance(b, int) and (
-        -32 < b < 32 or (is_integer_pow and b >= 0)
-    )
-    if embed_exponent:
-        loader = a.make_loader()
-
-        def fn(idx):
-            return pow_recursive(loader(idx), b, a.get_dtype())
-
-        return Pointwise.create(
-            device=a.get_device(),
-            dtype=a.get_dtype(),
-            inner_fn=fn,
-            ranges=a.get_size(),
-        )
-
-    if isinstance(a, Number):
-        if a == 1:
-            return full_like(b, 1)
-
-        if a == 2 and is_float_dtype(b.get_dtype()):
-            return exp2(b)
-
-    if is_integer_pow:
-        # ops.pow doesn't work for integers
-        if isinstance(a, Number):
-            return fallback_pow_scalar(a, b)
-        elif isinstance(b, Number):
-            return fallback_pow_tensor_scalar(a, b)
-        else:
-            return fallback_pow_tensor_tensor(a, b)
-
-    return pow_native(a, b)
-
-
-def mutate_to(changed, val, unsafe_alias=False):
-    if isinstance(changed, TensorBox):
-        changed_data = changed.data
-    else:
-        changed_data = changed
-    if isinstance(val, TensorBox):
-        val = val.data
-
-    if not isinstance(val, ir.StorageBox):
-        # introduce a copy to handle views
-        node = Pointwise.create(
-            device=changed.get_device(),
-            dtype=changed.get_dtype(),
-            inner_fn=val.make_loader(),
-            ranges=changed.get_size(),
-        )
-        assert isinstance(node, (BaseView, MutableBox))
-        val = node.data
-        assert isinstance(val, ir.StorageBox)
-
-    if isinstance(changed_data, ir.StorageBox) and not (
-        changed_data.is_input_buffer()
-        # In AOTI, module parameters and buffers are not lifted as graph inputs
-        or changed_data.is_module_buffer()
-        or isinstance(changed_data.data, ir.NopKernel)
-    ):
-        # Fast path, just swing the data pointer
-        val.realize()
-        changed_data.data = val.data
-        return changed
-
-    ir.MutationLayoutSHOULDREMOVE.realize_into(
-        val, changed_data, unsafe_alias=unsafe_alias
-    )
-    return changed
-
-
-@register_lowering(aten.fill_)
-def fill_(x, fill_value):
-    return mutate_to(x, full_like(x, fill_value))
-
-
-@register_lowering(aten.copy_, type_promotion_kind=None)
-def copy_(dst, src, non_blocking=False):
-    if dst is src:
-        # dst.copy_(dst) can happen from the reinplacing pass
-        return dst
-    src = to_device(src, dst.get_device())
-    src = to_dtype(src, dst.get_dtype())
-    src = expand(src, dst.get_size())
-    return mutate_to(dst, src)
-
-
-@make_pointwise
-def floordiv(a, b):
-    return ops.floordiv(a, b)
-
-
-@make_pointwise
-def truncdiv(a, b):
-    return ops.truncdiv(a, b)
-
-
-@make_pointwise
-def _div_rn(a, b):
-    return ops.div_rn(a, b)
-
-
-@register_lowering(aten.div, broadcast=True)
-def div_mode(a, b, rounding_mode=None):
-    both_integer = is_integer_type(a) and is_integer_type(b)
-    both_boolean = is_boolean_type(a) and is_boolean_type(b)
-
-    # floordiv and truncdiv need special handling for integer tensors on Triton,
-    # see the discussion at https://github.com/triton-lang/triton/issues/605
-    if rounding_mode == "floor":
-        assert not both_boolean, "floordiv operands can not be boolean at the same time"
-        # Use div_rn (IEEE round-to-nearest) instead of truediv here because
-        # Triton's default division uses an approximate reciprocal, which can
-        # produce a result slightly below the true quotient and cause floor()
-        # to round down by one.
-        return floordiv(a, b) if both_integer else floor(_div_rn(a, b))
-    if rounding_mode == "trunc":
-        assert not both_boolean, "truncdiv operands can not be boolean at the same time"
-        return truncdiv(a, b) if both_integer else trunc(div(a, b))
-    return div(a, b)
-
-
-@register_lowering([aten.mul], broadcast=True)
-def mul(a, b):
-    both_bool = is_boolean_type(a) and is_boolean_type(b)
-    if both_bool:
-        return logical_and(a, b)
-    else:
-        fn = ops_wrapper(aten.mul.__name__)
-        return make_pointwise(fn)(a, b)
-
-
-def get_constant_value(x: ir.IRNode) -> ir.Constant | None:
-    """Try convert an arbitrary IR node into an ir.Constant value"""
-
-    # First try unwrapping the IRNode to see if it is already an ir.Constant
-    # Optional step, but avoids unnecessary inner_fn evaluation.
-    if isinstance(x, ir.MutableBox):
-        return get_constant_value(x.data)
-    if isinstance(x, ir.BaseView):
-        return get_constant_value(x.unwrap_view())
-    if isinstance(x, ir.Constant):
-        return x
-
-    # If the unwrapped node is not an ir.Constant, try evaluating inner_fn
-    # to see if the returned value is from an `ops.constant` call
-    if not isinstance(x, ir.Loops):
-        return None
-
-    handler = torch._inductor.ops_handler.ExtractConstantsHandler(x.get_device())
-    with (
-        V.set_ops_handler(handler),
-        patch.object(ir.FlexibleLayout, "allow_indexing", True),
-    ):
-        out = x.inner_fn(*x.inner_fn_args())
-
-    assert isinstance(out, torch._inductor.virtualized.OpsValue)
-    if isinstance(out.value, ir.Constant):
-        return out.value
-    return None
-
-
-# NOTE: prims.div maps to a / b in C, so performs truncation division on
-#   integer inputs and true division for floating and complex inputs.
-@register_lowering([prims.div], broadcast=True)
-def div_prim(a, b):
-    is_integral = all(is_boolean_type(x) or is_integer_type(x) for x in [a, b])
-
-    if is_integral:
-        return truncdiv(a, b)
-
-    # Disable CPU optimization to avoid precision issues.
-    # see https://github.com/pytorch/pytorch/issues/157959
-    if (divisor := get_constant_value(b)) is not None and a.get_device().type != "cpu":
-        # Replace divide by constant with multiply by reciprocal
-
-        if divisor.value == 0:
-            reciprocal = math.copysign(float("inf"), divisor.value)
-        else:
-            reciprocal = 1.0 / divisor.value
-        return mul(a, reciprocal)
-
-    def fn(*args):
-        return ops.truediv(*args)
-
-    return make_pointwise(fn)(a, b)
-
-
-@register_lowering(
-    [aten.true_divide, aten.div.Tensor],
-    broadcast=True,
-    type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT,
-)
-def div(a, b):
-    a, b = promote_constants(
-        (a, b), type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT
-    )
-    return div_prim(a, b)
-
-
-@register_lowering([aten.fmod, prims.fmod], broadcast=True)
-def fmod(a, b):
-    is_integral = is_boolean_type(a) or is_integer_type(a)
-
-    if is_integral:
-
-        def fn(a, b):
-            return ops.mod(a, b)
-
-    else:
-
-        def fn(a, b):
-            return ops.fmod(a, b)
-
-    return make_pointwise(fn)(a, b)
-
-
-@register_lowering([aten.sum, prims.sum])
-def sum_(x, axis=None, keepdims=False, *, dtype=None):
-    if (
-        is_integer_dtype(x.get_dtype()) or is_boolean_dtype(x.get_dtype())
-    ) and dtype is None:
-        dtype = torch.int64
-
-    fn = make_reduction("sum", override_return_dtype=dtype)
-    return fn(x, axis, keepdims, dtype=dtype)
-
-
-fallback_cumsum = fallback_handler(aten.cumsum.default)
-fallback_cumprod = fallback_handler(aten.cumprod.default)
-fallback_logcumsumexp = fallback_handler(aten.logcumsumexp.default)
-fallback_cummax = fallback_handler(aten.cummax.default)
-fallback_cummin = fallback_handler(aten.cummin.default)
-
-
-@register_lowering(aten.cumsum)
-def cumsum(x, axis=None, dtype=None):
-    if (
-        is_integer_dtype(x.get_dtype()) or is_boolean_dtype(x.get_dtype())
-    ) and dtype is None:
-        dtype = torch.int64
-
-    if len(x.get_size()) == 0:
-        assert axis in [0, -1]
-        dtype = dtype or x.get_dtype()
-        return to_dtype(x, dtype, copy=True)
-
-    def combine_fn(a_tuple, b_tuple):
-        (a,) = a_tuple
-        (b,) = b_tuple
-        return (ops.add(a, b),)
-
-    kwargs = _make_scan_inner(x, axis=axis, dtype=dtype)
-    (result,) = ir.Scan.create(**kwargs, combine_fn=combine_fn)
-    if result is None:
-        return fallback_cumsum(x, dim=axis, dtype=dtype)
-    return result
-
-
-@register_lowering(aten.cumprod)
-def cumprod(x, axis=None, dtype=None):
-    if (
-        is_integer_dtype(x.get_dtype()) or is_boolean_dtype(x.get_dtype())
-    ) and dtype is None:
-        dtype = torch.int64
-
-    if len(x.get_size()) == 0:
-        assert axis in [0, -1]
-        dtype = dtype or x.get_dtype()
-        return to_dtype(x, dtype, copy=True)
-
-    def combine_fn(a_tuple, b_tuple):
-        (a,) = a_tuple
-        (b,) = b_tuple
-        return (ops.mul(a, b),)
-
-    kwargs = _make_scan_inner(x, axis=axis, dtype=dtype)
-    (result,) = ir.Scan.create(**kwargs, combine_fn=combine_fn)
-    if result is None:
-        return fallback_cumprod(x, dim=axis, dtype=dtype)
-    return result
-
-
-@register_lowering(aten.logcumsumexp)
-def logcumsumexp(x, dim):
-    def log_add_exp_helper(a_tuple, b_tuple):
-        (a,) = a_tuple
-        (b,) = b_tuple
-        min_v = ops.minimum(a, b)
-        max_v = ops.maximum(a, b)
-        mask = (min_v != max_v) | (~ops.isinf(min_v))
-        return (ops.where(mask, ops.log1p(ops.exp(min_v - max_v)) + max_v, a),)
-
-    dtype = x.get_dtype()
-    if len(x.get_size()) == 0:
-        assert dim in [0, -1]
-        return clone(x)
-
-    kwargs = _make_scan_inner(x, axis=dim, dtype=dtype)
-    (result,) = ir.Scan.create(**kwargs, combine_fn=log_add_exp_helper)
-    if result is None:
-        return fallback_logcumsumexp(x, dim=dim)
-    return result
-
-
-@register_lowering(aten.cummax, type_promotion_kind=None)
-def cummax(x, axis=None):
-    if len(x.get_size()) == 0:
-        assert axis in [0, -1]
-        return clone(x), empty_like(x, dtype=torch.int64)
-
-    dtype = x.get_dtype()
-    combine_fn = ir.get_reduction_combine_fn(
-        "argmax", dtype=dtype, arg_break_ties_left=False
-    )
-
-    kwargs = _make_scan_inner(x, axis=axis, dtype=dtype)
-    kwargs["dtypes"] = (dtype, torch.int64)
-    kwargs["inner_fns"] = (
-        x.make_loader(),
-        lambda idx: ops.index_expr(idx[axis], torch.int64),
-    )
-    values, indices = ir.Scan.create(**kwargs, combine_fn=combine_fn)  # type: ignore[arg-type]
-    if values is None:
-        return fallback_cummax(x, dim=axis)
-    return values, indices
-
-
-@register_lowering(aten.cummin, type_promotion_kind=None)
-def cummin(x, axis=None):
-    if len(x.get_size()) == 0:
-        assert axis in [0, -1]
-        return clone(x), empty_like(x, dtype=torch.int64)
-
-    dtype = x.get_dtype()
-    combine_fn = ir.get_reduction_combine_fn(
-        "argmin", dtype=dtype, arg_break_ties_left=False
-    )
-
-    kwargs = _make_scan_inner(x, axis=axis, dtype=dtype)
-    kwargs["dtypes"] = (dtype, torch.int64)
-    kwargs["inner_fns"] = (
-        x.make_loader(),
-        lambda idx: ops.index_expr(idx[axis], torch.int64),
-    )
-    values, indices = ir.Scan.create(**kwargs, combine_fn=combine_fn)  # type: ignore[arg-type]
-    if values is None:
-        return fallback_cummin(x, dim=axis)
-    return values, indices
-
-
-@register_lowering(aten.prod)
-def prod(x, axis=None, keepdims=False, *, dtype=None):
-    if (
-        is_integer_dtype(x.get_dtype()) or is_boolean_dtype(x.get_dtype())
-    ) and dtype is None:
-        dtype = torch.int64
-
-    fn = make_reduction("prod", override_return_dtype=dtype)
-    return fn(x, axis, keepdims, dtype=dtype)
-
-
-@register_lowering(aten.any)
-def reduce_any(x, dim=None, keepdim=False):
-    x = to_dtype(x, torch.bool)
-    return make_reduction("any")(x, axis=dim, keepdims=keepdim)
-
-
-@register_lowering(aten.max, type_promotion_kind=None)
-def reduce_max(x, dim=None, keepdim=False):
-    if dim is not None:
-        return (
-            reduce_amax(x, axis=dim, keepdims=keepdim),
-            reduce_argmax(x, axis=dim, keepdims=keepdim),
-        )
-
-    return reduce_amax(x, axis=None, keepdims=keepdim)
-
-
-@register_lowering(aten.min, type_promotion_kind=None)
-def reduce_min(x, dim=None, keepdim=False):
-    if dim is not None:
-        return (
-            reduce_amin(x, axis=dim, keepdims=keepdim),
-            reduce_argmin(x, axis=dim, keepdims=keepdim),
-        )
-
-    return reduce_amin(x, axis=None, keepdims=keepdim)
-
-
-register_lowering(prims.xor_sum)(make_reduction("xor_sum"))
-reduce_amax = register_lowering(aten.amax)(make_reduction("max"))
-reduce_amin = register_lowering(aten.amin)(make_reduction("min"))
-reduce_argmax = register_lowering(aten.argmax)(
-    make_reduction("argmax", override_return_dtype=torch.int64)
-)
-reduce_argmin = register_lowering(aten.argmin)(
-    make_reduction("argmin", override_return_dtype=torch.int64)
-)
-
-add = register_pointwise(
-    aten.add,
-    allow_alpha=True,
-    use_fma_for_alpha=True,
-    override_fn_when_input_bool="logical_or",
-)
 
 sort_fallback = fallback_handler(aten.sort.stable, add_to_fallback_set=False)
 
@@ -7325,443 +5130,6 @@ def sort(x, dim=-1, descending=False):
     return sort_stable(x, stable=False, dim=dim, descending=descending)
 
 
-def register_pointwise_numeric(op, name=None, triton_fallback=None):
-    return register_pointwise(
-        op,
-        name=name,
-        type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT,
-        triton_fallback=triton_fallback,
-    )
-
-
-def register_pointwise_numeric_ldf64(op: torch._ops.OpOverloadPacket):
-    register_op_requires_libdevice_fp64(op.__name__)
-    return register_pointwise(
-        op,
-        type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT,
-    )
-
-
-rsqrt = register_pointwise_numeric(aten.rsqrt)
-exp = register_pointwise_numeric_ldf64(aten.exp)
-exp2 = register_pointwise_numeric(aten.exp2)
-expm1 = register_pointwise_numeric(aten.expm1)
-relu = register_pointwise(aten.relu)
-sigmoid = register_pointwise_numeric_ldf64(aten.sigmoid)
-sqrt = register_pointwise_numeric_ldf64(aten.sqrt)
-square = register_pointwise(aten.square)
-sub = register_pointwise(aten.sub, allow_alpha=True)
-
-
-@register_lowering(aten.addcmul, broadcast=True)
-def addcmul(self, tensor1, tensor2, *, value=1):
-    """
-    Computes self + value * tensor1 * tensor2 using FMA for better precision.
-
-    Matches eager CUDA kernel order: self + value * (tensor1 * tensor2)
-    This is computed as: fma(value, tensor1 * tensor2, self)
-
-    Note: FMA is only used for floating-point types on non-AMD GPUs. For integer types,
-    we fall back to regular arithmetic since FMA doesn't support integers.
-
-    For floating-point types, we use mul_rn (round-to-nearest multiplication)
-    to force rounding of the product before the FMA. This prevents Triton's
-    compiler from fusing the multiplication with the FMA, matching eager's
-    rounding behavior.
-    """
-    dtype = get_promoted_dtype(
-        self,
-        tensor1,
-        tensor2,
-        type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.DEFAULT,
-    )
-
-    self_loader = self.make_loader()
-    t1_loader = tensor1.make_loader()
-    t2_loader = tensor2.make_loader()
-
-    # FMA/mul_rn/div_rn are only available for floating-point types on CUDA (non-AMD)
-    device = self.get_device()
-    use_fma = (
-        dtype.is_floating_point
-        and not torch.version.hip
-        and device is not None
-        and device.type in ["cuda", "xpu"]
-    )
-
-    def inner_fn(idx):
-        self_val = self_loader(idx)
-        t1_val = t1_loader(idx)
-        t2_val = t2_loader(idx)
-
-        if value == 1 and use_fma:
-            return ops.fma(t1_val, t2_val, self_val)
-
-        # Match eager order: self + value * (tensor1 * tensor2)
-        # Compute tensor1 * tensor2 first
-        if use_fma:
-            # Use mul_rn to force rounding of the product, preventing Triton
-            # from fusing t1*t2 with the subsequent FMA
-            t1_times_t2 = ops.mul_rn(t1_val, t2_val)
-        else:
-            t1_times_t2 = ops.mul(t1_val, t2_val)
-
-        # Use index_expr for sympy expressions (e.g., from .item()), constant otherwise
-        if isinstance(value, sympy.Basic):
-            value_expr = ops.index_expr(value, dtype)
-        else:
-            value_expr = ops.constant(value, dtype)
-
-        if use_fma:
-            # Use FMA for floating-point types for better precision
-            return ops.fma(value_expr, t1_times_t2, self_val)
-        else:
-            # Fall back to regular arithmetic for integer types
-            return ops.add(self_val, ops.mul(value_expr, t1_times_t2))
-
-    return Pointwise.create(
-        device=self.get_device(),
-        dtype=dtype,
-        inner_fn=inner_fn,
-        ranges=self.get_size(),
-    )
-
-
-@register_lowering(aten.addcdiv, broadcast=True)
-def addcdiv(self, tensor1, tensor2, *, value=1):
-    """
-    Computes self + value * (tensor1 / tensor2) using FMA for better precision.
-
-    Matches eager CUDA kernel order: self + value * (tensor1 / tensor2)
-    This is computed as: fma(value, tensor1 / tensor2, self)
-
-    For value=1: self + tensor1 / tensor2 (no FMA needed, just add the division)
-    For value!=1: fma(value, div_rn(tensor1, tensor2), self)
-
-    Note: FMA is only used for floating-point types on non-AMD GPUs. For integer types,
-    we fall back to regular arithmetic since FMA doesn't support integers.
-
-    We use div_rn (round-to-nearest division) to force proper rounding, preventing
-    Triton from fusing operations in ways that change the rounding behavior.
-    """
-    dtype = get_promoted_dtype(
-        self,
-        tensor1,
-        tensor2,
-        type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT,
-    )
-
-    self_loader = self.make_loader()
-    t1_loader = tensor1.make_loader()
-    t2_loader = tensor2.make_loader()
-
-    # FMA/mul_rn/div_rn are only available for floating-point types on CUDA (non-AMD)
-    device = self.get_device()
-    use_fma = (
-        dtype.is_floating_point
-        and not torch.version.hip
-        and device is not None
-        and device.type in ["cuda", "xpu"]
-    )
-
-    def inner_fn(idx):
-        self_val = self_loader(idx)
-        t1_val = t1_loader(idx)
-        t2_val = t2_loader(idx)
-
-        # Compute tensor1 / tensor2 first
-        # Use div_rn for round-to-nearest division on CUDA to match eager behavior
-        if use_fma:
-            t1_div_t2 = ops.div_rn(t1_val, t2_val)
-        else:
-            t1_div_t2 = ops.truediv(t1_val, t2_val)
-
-        if value == 1:
-            # For value=1, just add the division result (no FMA needed)
-            return ops.add(self_val, t1_div_t2)
-
-        # Use index_expr for sympy expressions (e.g., from .item()), constant otherwise
-        if isinstance(value, sympy.Basic):
-            value_expr = ops.index_expr(value, dtype)
-        else:
-            value_expr = ops.constant(value, dtype)
-
-        if use_fma:
-            # Use FMA for floating-point types for better precision
-            return ops.fma(value_expr, t1_div_t2, self_val)
-        else:
-            # Fall back to regular arithmetic for integer types
-            return ops.add(self_val, ops.mul(value_expr, t1_div_t2))
-
-    return Pointwise.create(
-        device=self.get_device(),
-        dtype=dtype,
-        inner_fn=inner_fn,
-        ranges=self.get_size(),
-    )
-
-
-_foreach_addcmul_scalar = register_foreach_pointwise(
-    aten._foreach_addcmul.Scalar, addcmul, allow_alpha=True, scalar_kwarg="value"
-)
-_foreach_addcdiv_scalar = register_foreach_pointwise(
-    aten._foreach_addcdiv.Scalar, addcdiv, allow_alpha=True, scalar_kwarg="value"
-)
-
-
-register_pointwise_numeric_ldf64(aten.cos)
-register_pointwise_numeric_ldf64(aten.sin)
-abs = register_pointwise(aten.abs)
-bitwise_and = register_pointwise(aten.bitwise_and)
-bitwise_left_shift = register_pointwise(aten.bitwise_left_shift)
-bitwise_not = register_pointwise(
-    aten.bitwise_not, override_fn_when_input_bool="logical_not"
-)
-bitwise_or = register_pointwise(aten.bitwise_or)
-bitwise_right_shift = register_pointwise(aten.bitwise_right_shift)
-bitwise_xor = register_pointwise(aten.bitwise_xor)
-register_pointwise_numeric(aten.lgamma)
-erf = register_pointwise_numeric(aten.erf)
-register_lowering(
-    aten.special_erf, type_promotion_kind=ELEMENTWISE_TYPE_PROMOTION_KIND.INT_TO_FLOAT
-)(erf)
-
-register_pointwise_numeric(aten.log1p)
-register_pointwise_numeric(aten.tan)
-register_pointwise_numeric(aten.tanh)
-register_pointwise_numeric_ldf64(aten.log)
-logical_and = register_pointwise(
-    aten.logical_and,
-    type_promotion_kind=None,
-    convert_input_to_bool=True,
-    override_return_dtype=torch.bool,
-)
-logical_not = register_pointwise(
-    aten.logical_not,
-    type_promotion_kind=None,
-    convert_input_to_bool=True,
-    override_return_dtype=torch.bool,
-)
-logical_or = register_pointwise(
-    aten.logical_or,
-    type_promotion_kind=None,
-    convert_input_to_bool=True,
-    override_return_dtype=torch.bool,
-)
-logical_xor = register_pointwise(
-    aten.logical_xor,
-    type_promotion_kind=None,
-    convert_input_to_bool=True,
-    override_return_dtype=torch.bool,
-)
-maximum = register_pointwise(aten.maximum)
-minimum = register_pointwise(aten.minimum)
-register_lowering(aten.clamp_min)(maximum)
-register_lowering(aten.clamp_max)(minimum)
-neg = register_pointwise(aten.neg)
-abs = register_pointwise(aten.abs)
-reciprocal = register_pointwise_numeric(aten.reciprocal)
-register_pointwise(aten.remainder)
-sign = register_pointwise(aten.sign, override_fn_when_input_bool="identity")
-register_pointwise(aten.ceil)
-register_pointwise(aten.signbit, override_return_dtype=torch.bool)
-
-register_lowering(aten._neg_view)(neg)
-
-register_pointwise(aten.le, override_return_dtype=torch.bool)
-register_pointwise(aten.lt, override_return_dtype=torch.bool)
-register_pointwise(aten.ge, override_return_dtype=torch.bool)
-gt = register_pointwise(aten.gt, override_return_dtype=torch.bool)
-register_pointwise(aten.eq, override_return_dtype=torch.bool)
-register_pointwise(aten.ne, override_return_dtype=torch.bool)
-
-register_pointwise_numeric(aten.cosh)
-register_pointwise_numeric(aten.sinh)
-register_pointwise_numeric(aten.acos)
-register_pointwise_numeric(aten.acosh)
-register_pointwise_numeric(aten.asin)
-register_pointwise_numeric(aten.asinh)
-register_pointwise_numeric(aten.atan2)
-register_pointwise_numeric(aten.atan)
-register_pointwise_numeric(aten.atanh)
-register_pointwise_numeric(aten.copysign)
-register_pointwise_numeric(aten.erfc)
-register_pointwise_numeric(aten.erfinv)
-register_pointwise_numeric(aten.hypot)
-register_pointwise_numeric(aten.log10)
-register_pointwise_numeric(aten.log2)
-register_pointwise_numeric(aten.nextafter)
-
-from .codegen.common import BackendFeature, pointwise_overrides_data
-
-
-def _get_pointwise_overrides(ns, name):
-    data = pointwise_overrides_data[name]
-    op = getattr(ns, data.name, None)
-    if op is None:
-        return
-
-    def make_triton_fallback(op):
-        if data.triton is None:
-            return fallback_handler(op)
-
-    if isinstance(op, torch._ops.OpOverloadPacket):
-        for olname in op.overloads():
-            ol = getattr(op, olname)
-            yield ol, data.type_promotion_kind, make_triton_fallback(ol)
-    else:
-        yield op, data.type_promotion_kind, make_triton_fallback(op)
-
-
-for name in pointwise_overrides_data:
-    for op, type_promotion_kind, triton_fallback in _get_pointwise_overrides(
-        aten, name
-    ):
-        register_pointwise(
-            op,
-            name=name,
-            type_promotion_kind=type_promotion_kind,
-            triton_fallback=triton_fallback,
-        )
-
-    for op, type_promotion_kind, triton_fallback in _get_pointwise_overrides(
-        prims, name
-    ):
-        register_pointwise(
-            op,
-            name=name,
-            type_promotion_kind=type_promotion_kind,
-            triton_fallback=triton_fallback,
-        )
-
-
-foreach_add_list = register_foreach_pointwise(
-    aten._foreach_add.List, add, allow_alpha=True
-)
-foreach_add_scalar = register_foreach_pointwise(
-    aten._foreach_add.Scalar, add, allow_alpha=True
-)
-register_foreach_pointwise(aten._foreach_add.Tensor, add, allow_alpha=True)
-foreach_mul_list = register_foreach_pointwise(aten._foreach_mul.List, mul)
-register_foreach_pointwise(aten._foreach_mul.Tensor, mul)
-foreach_mul_scalar = register_foreach_pointwise(aten._foreach_mul.Scalar, mul)
-register_foreach_pointwise(aten._foreach_sub.List, sub)
-register_foreach_pointwise(aten._foreach_sub.Scalar, sub)
-register_foreach_pointwise(aten._foreach_neg.default, neg)
-register_foreach_pointwise(aten._foreach_abs.default, abs)
-register_foreach_pointwise(aten._foreach_pow.Scalar, pow)
-register_foreach_pointwise(aten._foreach_pow.List, pow)
-register_foreach_pointwise(aten._foreach_pow.ScalarAndTensor, pow)
-foreach_div_list = register_foreach_pointwise(aten._foreach_div.List, div)
-register_foreach_pointwise(aten._foreach_div.Tensor, div)
-foreach_div_scalar = register_foreach_pointwise(aten._foreach_div.Scalar, div)
-register_foreach_pointwise(aten._foreach_sqrt, sqrt)
-register_foreach_pointwise(aten._foreach_rsqrt, rsqrt)
-register_foreach_pointwise(aten._foreach_maximum.List, maximum)
-register_foreach_pointwise(aten._foreach_maximum.Scalar, maximum)
-register_foreach_pointwise(aten._foreach_minimum.List, minimum)
-register_foreach_pointwise(aten._foreach_minimum.Scalar, minimum)
-register_foreach_pointwise(aten._foreach_clamp_min.List, maximum)
-register_foreach_pointwise(aten._foreach_clamp_min.Scalar, maximum)
-register_foreach_pointwise(aten._foreach_clamp_max.List, minimum)
-register_foreach_pointwise(aten._foreach_clamp_max.Scalar, minimum)
-register_foreach_pointwise(aten._foreach_reciprocal, reciprocal)
-register_foreach_pointwise(aten._foreach_sign, sign)
-register_foreach_pointwise(aten._foreach_clone, clone)
-foreach_copy = register_foreach_pointwise(aten._foreach_copy, copy)
-
-
-# these are only encountered as outputs of the graph
-# reinplacing epilogue copies improves compile time
-# by removing extra buffers sent to the scheduler.
-def register_foreach_inplace(aten_op, outplace_aten_op, outplace_op):
-    inplaceable_foreach_ops[outplace_aten_op] = aten_op
-    inplace_foreach_ops.add(aten_op)
-
-    def fn(*args, **kwargs):
-        results = outplace_op(*args, **kwargs)
-        mut_results = []
-        for arg, result in zip(args[0], results):
-            mut_results.append(mutate_to(arg, result, unsafe_alias=True))
-
-        return mut_results
-
-    _register_foreach_lowering(aten_op, fn)
-
-
-register_foreach_inplace(
-    aten._foreach_add_.List, aten._foreach_add.List, foreach_add_list
-)
-register_foreach_inplace(
-    aten._foreach_add_.Scalar, aten._foreach_add.Scalar, foreach_add_scalar
-)
-register_foreach_inplace(
-    aten._foreach_mul_.List, aten._foreach_mul.List, foreach_mul_list
-)
-register_foreach_inplace(
-    aten._foreach_mul_.Scalar, aten._foreach_mul.Scalar, foreach_mul_scalar
-)
-register_foreach_inplace(
-    aten._foreach_div_.List, aten._foreach_div.List, foreach_div_list
-)
-register_foreach_inplace(
-    aten._foreach_div_.Scalar, aten._foreach_div.Scalar, foreach_div_scalar
-)
-register_foreach_inplace(
-    aten._foreach_copy_.default, aten._foreach_copy.default, foreach_copy
-)
-register_foreach_inplace(
-    aten._foreach_addcmul_.Scalar,
-    aten._foreach_addcmul.Scalar,
-    _foreach_addcmul_scalar,
-)
-register_foreach_inplace(
-    aten._foreach_addcdiv_.Scalar,
-    aten._foreach_addcdiv.Scalar,
-    _foreach_addcdiv_scalar,
-)
-
-
-def register_inplace(aten_op, outplace_op):
-    @register_lowering(aten_op, type_promotion_kind=None)
-    def fn(*args, **kwargs):
-        result = outplace_op(*args, **kwargs)
-        result = to_dtype(result, args[0].get_dtype())
-        return mutate_to(args[0], result)
-
-    return fn
-
-
-register_inplace(aten.add_, add)
-register_inplace(aten.bitwise_and_, bitwise_and)
-register_inplace(aten.bitwise_left_shift_, bitwise_left_shift)
-register_inplace(aten.bitwise_not_, bitwise_not)
-register_inplace(aten.bitwise_or_, bitwise_or)
-register_inplace(aten.bitwise_right_shift_, bitwise_right_shift)
-register_inplace(aten.bitwise_xor_, bitwise_xor)
-register_inplace(aten.mul_, mul)
-register_inplace(aten.div_.Tensor, div)
-register_inplace(aten.div_.Tensor_mode, div_mode)
-register_inplace(aten.logical_and_, logical_and)
-register_inplace(aten.logical_not_, logical_not)
-register_inplace(aten.logical_or_, logical_or)
-register_inplace(aten.logical_xor_, logical_xor)
-register_inplace(aten.sub_, sub)
-register_inplace(aten.relu_, relu)
-register_inplace(aten.sigmoid_, sigmoid)
-
-
-register_lowering(aten.__and__)(bitwise_and)
-register_lowering(aten.__lshift__)(bitwise_left_shift)
-register_lowering(aten.__or__)(bitwise_or)
-register_lowering(aten.__rshift__)(bitwise_right_shift)
-register_lowering(aten.__xor__)(bitwise_xor)
-
-register_inplace(aten.__iand__, aten.__and__)
-register_inplace(aten.__ilshift__, aten.__lshift__)
-register_inplace(aten.__ior__, aten.__or__)
-register_inplace(aten.__irshift__, aten.__rshift__)
-register_inplace(aten.__ixor__, aten.__xor__)
 
 
 @register_lowering(aten.sym_constrain_range)
@@ -8380,6 +5748,24 @@ def lower_inline_asm_elementwise(
         inner_fn=inner_fn,
         ranges=list(inputs[0].get_size()),
     )
+
+
+def _reexport_lowering_symbols(module) -> None:
+    # Keep existing torch._inductor.lowering attribute lookups working
+    # while the implementation lives in smaller category modules.
+    for name in module.__all__:
+        globals()[name] = getattr(module, name)
+
+
+from . import elementwise_lowerings, pooling_lowerings, reduction_lowerings
+
+
+for lowering_module in (
+    elementwise_lowerings,
+    reduction_lowerings,
+    pooling_lowerings,
+):
+    _reexport_lowering_symbols(lowering_module)
 
 
 # populate lowerings defined in kernel/*
